+++ src/Michalmanko/Holiday/HolidayFactory.php
@@ -1,5 +1,14 @@
 <?php
 
+/*
+ * This file is part of the Holiday Library.
+ *
+ * (c) Michał Mańko <github@michalmanko.com>
+ *
+ * For the full copyright and license information, please view the LICENSE.md
+ * file that was distributed with this source code.
+ */
+
 namespace Michalmanko\Holiday;
 
 use DateTimeZone;
@@ -40,7 +49,7 @@ abstract class HolidayFactory
      */
     public static function registerProvider($countryCode, $providerClassName)
     {
-        $countryCode       = \strtoupper($countryCode);
+        $countryCode       = strtoupper($countryCode);
         $providerClassName = (string) $providerClassName;
 
         static::$providers[$countryCode] = $providerClassName;
@@ -49,19 +58,19 @@ abstract class HolidayFactory
     /**
      * Unregisters a holidays provider class.
      *
-     * @param string $countryCodeOrProviderClassName Country code of the provider or the provider class name
+     * @param string $providerClassName Country code of the provider or the provider class name
      *
-     * @return boolean
+     * @return bool
      */
-    public static function unregisterProvider($countryCodeOrProviderClassName)
+    public static function unregisterProvider($providerClassName)
     {
-        if (\array_key_exists($countryCodeOrProviderClassName, static::$providers)) {
-            unset(static::$providers[$countryCodeOrProviderClassName]);
+        if (array_key_exists($providerClassName, static::$providers)) {
+            unset(static::$providers[$providerClassName]);
 
             return true;
         }
 
-        $index = \array_search((string) $countryCodeOrProviderClassName, static::$providers, true);
+        $index = array_search((string) $providerClassName, static::$providers, true);
         if ($index !== false) {
             unset(static::$providers[$index]);
 
@@ -84,11 +93,11 @@ abstract class HolidayFactory
     public static function createProvider($countryCode, DateTimeZone $timezone = null)
     {
         // Use the supplied provider class
-        if (\substr($countryCode, 0, 1) == '\\') {
+        if (substr($countryCode, 0, 1) == '\\') {
             $className = $countryCode;
-        } elseif (\array_key_exists(\strtoupper($countryCode), static::$providers)) {
-            $countryCode = \strtoupper($countryCode);
-            if (\substr(static::$providers[$countryCode], 0, 1) == '\\') {
+        } elseif (array_key_exists(strtoupper($countryCode), static::$providers)) {
+            $countryCode = strtoupper($countryCode);
+            if (substr(static::$providers[$countryCode], 0, 1) == '\\') {
                 $className = static::$providers[$countryCode];
             } else {
                 $className = '\\' . __NAMESPACE__ . '\\Provider\\' . static::$providers[$countryCode];
@@ -97,8 +106,8 @@ abstract class HolidayFactory
             $className = '\\' . __NAMESPACE__ . '\\Provider\\' . $countryCode;
         }
 
-        if (!\class_exists($className)) {
-            throw new InvalidArgumentException(\sprintf('Cannot find Holiday provider class "%s"', $className));
+        if (!class_exists($className)) {
+            throw new InvalidArgumentException(sprintf('Cannot find Holiday provider class "%s"', $className));
         }
 
         $provider = new $className($timezone);