+++ src/Michalmanko/Holiday/Provider/Denmark.php
@@ -1,10 +1,22 @@
 <?php
 
+/*
+ * This file is part of the Holiday Library.
+ *
+ * (c) Michał Mańko <github@michalmanko.com>
+ *
+ * For the full copyright and license information, please view the LICENSE.md
+ * file that was distributed with this source code.
+ */
+
 namespace Michalmanko\Holiday\Provider;
 
 use ArrayObject;
 use Michalmanko\Holiday\Holiday;
 
+/**
+ * Danish Holidays Provider.
+ */
 class Denmark extends AbstractProvider
 {
     /**