+++ src/Michalmanko/Holiday/Provider/Exception/InvalidArgumentException.php
@@ -1,6 +1,13 @@
 <?php
 
-
+/*
+ * This file is part of the Holiday Library.
+ *
+ * (c) Michał Mańko <github@michalmanko.com>
+ *
+ * For the full copyright and license information, please view the LICENSE.md
+ * file that was distributed with this source code.
+ */
 
 namespace Michalmanko\Holiday\Provider\Exception;