+++ src/Michalmanko/Holiday/Test/Provider/NotProviderTest.php
@@ -1,9 +1,0 @@
-<?php
-
-
-
-namespace Michalmanko\Holiday\Test\Provider;
-
-class NotProviderTest
-{
-}