+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 CHANGELOG
 =========
 
+1.0.0
+----------
+ + upgrade phpunit to version 4.6
+ + upgrade php-cs-fixer to version 1.8
+ + moving tests from ./src to ./tests
+ + start supporting PHP 7
+
 0.2.0
 ----------
  + Danish Holidays (thanks to @lsv)